refactor: components 폴더 재구성
