+++ src/components/common/SkeletonCard.tsx
