+++ src/pages/MainPage.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import SearchBar from '@/components/SearchBar';
 import ProfileCard from '@/components/ProfileCard';
-import SeoHelmet from '@/components/SeoHelmet';
+import SeoHelmet from '../components/shared/SeoHelmet';
 
 interface MainPageProps {}
